# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/dealerships": {
      "post": {
        "operationId": "DealershipsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDealershipDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Concessionária criada.",
            "headers": {
              "Location": {
                "description": "URL do recurso criado",
                "schema": {
                  "type": "string",
                  "example": "/api/customers/{id}"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Dealership"
                }
              }
            }
          },
          "400": {
            "description": "Entrada inválida (validação do DTO, UUID malformado ou campo não permitido).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente, inválido ou expirado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Papel sem permissão, ou recurso de outro cliente.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Limite de requisições excedido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Cria uma concessionária (admin)",
        "tags": [
          "dealerships"
        ]
      },
      "get": {
        "operationId": "DealershipsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Dealership"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Token ausente, inválido ou expirado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Papel sem permissão, ou recurso de outro cliente.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Limite de requisições excedido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Lista concessionárias (qualquer papel)",
        "tags": [
          "dealerships"
        ]
      }
    },
    "/api/dealerships/{id}": {
      "get": {
        "operationId": "DealershipsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Dealership"
                }
              }
            }
          },
          "400": {
            "description": "Entrada inválida (validação do DTO, UUID malformado ou campo não permitido).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente, inválido ou expirado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Papel sem permissão, ou recurso de outro cliente.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Recurso não encontrado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Limite de requisições excedido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Detalha uma concessionária (qualquer papel)",
        "tags": [
          "dealerships"
        ]
      },
      "patch": {
        "operationId": "DealershipsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDealershipDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Dealership"
                }
              }
            }
          },
          "400": {
            "description": "Entrada inválida (validação do DTO, UUID malformado ou campo não permitido).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente, inválido ou expirado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Papel sem permissão, ou recurso de outro cliente.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Recurso não encontrado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Limite de requisições excedido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Atualiza parcialmente uma concessionária (admin)",
        "tags": [
          "dealerships"
        ]
      },
      "delete": {
        "operationId": "DealershipsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Concessionária removida."
          },
          "400": {
            "description": "Entrada inválida (validação do DTO, UUID malformado ou campo não permitido).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente, inválido ou expirado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Papel sem permissão, ou recurso de outro cliente.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Recurso não encontrado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Limite de requisições excedido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Remove uma concessionária (admin)",
        "tags": [
          "dealerships"
        ]
      }
    },
...
    "/api/customers/{customerId}/predictions": {
      "get": {
        "operationId": "CustomerPredictionsController_findByCustomer",
        "parameters": [
          {
            "name": "customerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PredictionResponseDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Entrada inválida (validação do DTO, UUID malformado ou campo não permitido).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente, inválido ou expirado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Papel sem permissão, ou recurso de outro cliente.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Recurso não encontrado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Limite de requisições excedido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Histórico de predições de um cliente (admin, analyst)",
        "tags": [
          "predictions"
        ]
      }
    }
  },
  "info": {
    "title": "Ford Conecta — API (Web Services)",
    "description": "Camada de Serviços REST da plataforma Ford Conecta — FIAP Ford 2026, Desafio 02.\n\n**Autenticação:** faça `POST /api/auth/login`, copie o `accessToken` e clique em **Authorize**. Rotas sem cadeado são públicas.\n\n**Papéis:** `admin` (tudo), `analyst` (pós-venda da concessionária) e `customer` (dono do veículo, só os próprios dados).\n\n**Erros:** toda resposta de erro segue o schema `ErrorResponseDto`.",
    "version": "2.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 404
          },
          "error": {
            "type": "string",
            "example": "Not Found"
          },
          "message": {
            "oneOf": [
              {
                "type": "string",
                "example": "Cliente não encontrado"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "example": [
                  "email must be an email"
                ]
              }
            ],
            "description": "Mensagem, ou lista de mensagens quando a validação de entrada falha"
          },
          "path": {
            "type": "string",
            "example": "/api/customers/00000000-0000-0000-0000-000000000000"
          },
          "timestamp": {
            "type": "string",
            "example": "2026-09-26T12:00:00.000Z"
          }
        },
        "required": [
          "statusCode",
          "error",
          "message",
          "path",
          "timestamp"
        ]
      },
...
      "Customer": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "document": {
            "type": "string",
            "description": "CPF (somente dígitos). Identificador único do cliente."
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "vehicles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Vehicle"
            }
          },
          "id": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "name",
          "document",
          "email",
          "phone",
          "city",
          "state",
          "vehicles",
          "id",
          "createdAt",
          "updatedAt"
        ]
      },
      "Dealership": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "region": {
            "type": "string"
          },
          "vehicles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Vehicle"
            }
          },
          "id": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "name",
          "city",
          "region",
          "vehicles",
          "id",
          "createdAt",
          "updatedAt"
        ]
      },
      "Vehicle": {
        "type": "object",
        "properties": {
          "vin": {
            "type": "string",
            "description": "Vehicle Identification Number — 17 caracteres."
          },
          "model": {
            "type": "string"
          },
          "modelYear": {
            "type": "number"
          },
          "purchaseDate": {
            "type": "string"
          },
          "odometer": {
            "type": "number",
            "description": "Hodômetro atual (km) — dado de veículo conectado (IoT). NÃO é usado na predição."
          },
          "customerId": {
            "type": "string"
          },
          "customer": {
            "$ref": "#/components/schemas/Customer"
          },
          "dealershipId": {
            "type": "string"
          },
          "dealership": {
            "$ref": "#/components/schemas/Dealership"
          },
          "id": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "vin",
          "model",
          "modelYear",
          "purchaseDate",
          "odometer",
          "customerId",
          "customer",
          "dealershipId",
          "dealership",
          "id",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "CreateDealershipDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 140,
            "example": "Ford Center Norte"
          },
          "city": {
            "type": "string",
            "maxLength": 80,
            "example": "São Paulo"
          },
          "region": {
            "type": "string",
            "maxLength": 40,
            "example": "Sudeste"
          }
        },
        "required": [
          "name",
          "city",
          "region"
        ]
      },
      "UpdateDealershipDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 140,
            "example": "Ford Center Norte"
          },
          "city": {
            "type": "string",
            "maxLength": 80,
            "example": "São Paulo"
          },
          "region": {
            "type": "string",
            "maxLength": 40,
            "example": "Sudeste"
          }
        }
      },
...
      "PredictionResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "customerId": {
            "type": "string",
            "format": "uuid"
          },
          "vehicleId": {
            "type": "string",
            "nullable": true,
            "format": "uuid"
          },
          "segment": {
            "enum": [
              "fiel",
              "abandono",
              "esquecido",
              "economico"
            ],
            "type": "string",
            "example": "abandono"
          },
          "evasionRisk": {
            "type": "number",
            "example": 0.9,
            "description": "Probabilidade estimada de evasão (0..1)"
          },
          "confidence": {
            "type": "number",
            "example": 0.41,
            "description": "Confiança na classificação (0..1)"
          },
          "recommendedAction": {
            "type": "string",
            "example": "Oferta agressiva na 1ª revisão (preço-âncora) + contato proativo..."
          },
          "modelVersion": {
            "type": "string",
            "example": "heuristic-v1"
          },
          "features": {
            "type": "object",
            "additionalProperties": true,
            "description": "Features do momento da compra usadas na predição"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "customerId",
          "vehicleId",
          "segment",
          "evasionRisk",
          "confidence",
          "recommendedAction",
          "modelVersion",
          "features",
          "createdAt"
        ]
      }
    }
  }
}